restructured classes
