+++ core/Addon.php
@@ -1,0 +1,8 @@
+<?php
+namespace sbnc\core;
+
+abstract class Addon extends Component
+{
+
+
+}
