+++ core/Util.php
@@ -1,0 +1,7 @@
+<?php
+namespace sbnc\core;
+
+abstract class Util extends Component
+{
+
+}
