+++ modules/Validate.php
@@ -2,6 +2,7 @@
 namespace sbnc\modules;
 
 use sbnc\Sbnc;
+use sbnc\core\Module;
 
 class Validate extends Module implements ModuleInterface
 {
